fix: resolve e2e build errors by removing orphaned TanStack Router files
- Moved unused src/routes/ directory to .old-routes/ (files were using TanStack Router which isn't installed)
- Added missing @types/react-router-dom dependency for proper TypeScript support
- Excluded orphaned routes from TypeScript compilation in tsconfig.app.json
- The actual routing is handled correctly via React Router v6 in src/App.tsx
