+++ .old-routes/routes/index.tsx
@@ -1,10 +1,16 @@
-import { createFileRoute } from '@tanstack/react-router'
-import { Suspense, lazy } from 'react'
+import { createFileRoute } from '@tanstack/react-router';
+import { Suspense, lazy } from 'react';
 
-const LoginPage = lazy(() => import('@/components/features/auth/login-page'))
+// Lazy load components for better performance
+const Home = lazy(() => import('@/components/common/layout').then((m) => ({ default: m.Home })));
 
+// Minimal loading fallback for fast FCP
 const PageSkeleton = () => (
-  <div className="min-h-screen bg-background flex flex-col" role="status" aria-label="Loading content">
+  <div
+    className="min-h-screen bg-background flex flex-col"
+    role="status"
+    aria-label="Loading content"
+  >
     <header className="border-b">
       <div className="flex h-16 items-center px-4 max-w-7xl mx-auto">
         <div className="text-xl font-bold">contributor.info</div>
@@ -20,14 +26,14 @@ const PageSkeleton = () => (
     </main>
     <span className="sr-only">Loading content, please wait...</span>
   </div>
-)
+);
 
-export const Route = createFileRoute('/login')({
-  // Disable SSR for login page since it requires browser APIs and user interaction
-  ssr: false,
+export const Route = createFileRoute('/')({
+  // Enable full SSR for the home page for better SEO and LCP
+  ssr: true,
   component: () => (
     <Suspense fallback={<PageSkeleton />}>
-      <LoginPage />
+      <Home />
     </Suspense>
   ),
-})
+});
