+++ package.json
@@ -218,6 +218,7 @@
     "@types/papaparse": "^5.5.0",
     "@types/react": "^19.2.7",
     "@types/react-dom": "^19.2.3",
+    "@types/react-router-dom": "^5.3.3",
     "@vitest/coverage-v8": "^3.1.1",
     "@vitest/ui": "^3.2.4",
     "concurrently": "^9.1.2",
